multimodal-rag-cli/requirements.txt
haystack-ai
google-genai-haystack
python-dotenv
pypdf
